Cleanup. Moved default AuditData implementation to another class
